backend/pyproject.toml
[project]
name = "aegistriage-backend"
version = "0.1.0"
description = "Backend for AegisTriage 2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.30.0",
    "supabase>=2.0.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "numpy>=2.0.0",
    "joblib>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

